Encode workout selection algorithm
This set of code is the encoding of the weighted random selection
algorithm the data team decided on for selecting games for workouts. It
also includes the utilities for calculating the weights for each game
including historical game plays and game weights.

Of note, our weighting algorithm tries to select games that a user would
be most likely to play instead of trying to maximize distribution over
our game list.